refactor code to a single "ret.go" file
